Adding styles, finishing task 4
Adding styles to the div so that they will arrange into required "boxes";

Pay attention the "div img" extension. I put there height and width of 100%, because the text, that currently replaces the pictures was not supposed to be the div boxes. If you define the images, then either remove the text in div boxes or modify height property of "div img";

W3C Validation did not show any errors;